requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
scikit-learn==1.7.2
imbalanced-learn==0.14.1
pandas==2.2.3
numpy==1.26.4
joblib==1.3.2
pydantic==2.12.5
python-dotenv==1.2.2
psycopg2-binary==2.9.11
sqlalchemy==2.0.48
scipy==1.13.1
httpx==0.28.1
pytest==9.0.2
pytest-cov==7.1.0
python-multipart==0.0.9